Rename symbols to match WordPress Coding standards
